feat: Revamp UI and demo structure
Key changes:

- Modularized example demos:
  - Each demo is now in its own file under `src/examples/demos/`.
  - `src/examples/pages.js` dynamically imports and lists all demos.
  - This simplifies adding, removing, or modifying individual demos.

- Revamped HUD for usability:
  - Implemented progressive disclosure for HUD controls.
  - Main HUD elements (camera mode, shortcuts, layout settings) are now in a popup menu, accessible via a new main menu button (☰).
  - This declutters the main view and focuses on content.
  - `hudSelectionInfo` remains directly visible.
  - Popup menu is dismissible via click-outside or Escape key.

- Code Quality & Enhancements:
  - Added JSDoc comments to `HudManager.js` for better documentation.
  - Refined CSS for new HUD elements for a clean and consistent look, including theme support.
  - Addressed minor FIXMEs from original demo code during refactoring.
  - Ensured `HudManager` can dynamically update camera mode options.

These changes improve the library's organization, user interface clarity,
and maintainability, preparing it for better distribution and future growth.
