+++ src/index.css
@@ -465,7 +465,8 @@ body {
     font-size: 13px;
 }
 
-#hud-layer > div,
+/* Common style for direct children of hud-layer that are individual items */
+#hud-layer > div:not(#hud-popup-menu),
 #hud-layer > select {
     position: absolute;
     background-color: rgba(var(--sg-bg-ui-rgb, 20, 20, 40), 0.7);
@@ -476,54 +477,156 @@ body {
     box-shadow: 0 1px 5px rgba(0,0,0,0.2);
     pointer-events: auto;
     white-space: nowrap;
-    transition: opacity var(--sg-transition-fast), background-color var(--sg-transition-fast);
+    transition: opacity var(--sg-transition-fast), background-color var(--sg-transition-fast), border-color var(--sg-transition-fast);
 }
 
-#hud-layer > div:hover,
+#hud-layer > div:not(#hud-popup-menu):hover,
 #hud-layer > select:hover {
     opacity: 1;
     background-color: rgba(var(--sg-bg-ui-rgb, 20, 20, 40), 0.85);
     border-color: rgba(var(--sg-accent-primary-rgb, 0, 208, 255), 0.7);
 }
 
-#hud-mode-indicator {
-    bottom: 10px;
+/* Main Menu Button */
+#hud-main-menu-button {
+    top: 10px;
     left: 10px;
+    font-size: 1.5em; /* Larger for easier clicking */
+    padding: 2px 10px; /* Adjusted padding */
+    cursor: pointer;
+    user-select: none;
+}
+#hud-main-menu-button.active {
+    background-color: rgba(var(--sg-accent-primary-rgb, 0, 208, 255), 0.2);
+    border-color: rgba(var(--sg-accent-primary-rgb, 0, 208, 255), 0.9);
+    opacity: 1;
 }
 
+
+/* Popup Menu */
+#hud-popup-menu {
+    position: absolute;
+    top: 50px; /* Below main menu button */
+    left: 10px;
+    background-color: rgba(var(--sg-bg-ui-rgb, 25, 25, 45), 0.9); /* Slightly different bg */
+    border: 1px solid var(--sg-accent-primary);
+    border-radius: 6px;
+    padding: 10px;
+    box-shadow: 0 3px 15px rgba(0,0,0,0.3);
+    pointer-events: auto;
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+    min-width: 220px;
+    opacity: 1;
+    transition: opacity 0.2s ease-out, transform 0.2s ease-out;
+    transform-origin: top left;
+}
+
+#hud-popup-menu.hidden {
+    opacity: 0;
+    transform: scale(0.95) translateY(-10px);
+    pointer-events: none;
+}
+
+.hud-menu-group {
+    display: flex;
+    flex-direction: column; /* Stack label and control */
+    gap: 4px;
+}
+.hud-menu-group label {
+    font-size: 0.9em;
+    color: var(--sg-text-secondary);
+    margin-bottom: 2px;
+}
+
+/* Styling for controls inside the popup menu */
+#hud-popup-menu select,
+#hud-popup-menu button {
+    width: 100%;
+    padding: 8px 10px;
+    background-color: var(--sg-node-button-bg);
+    color: var(--sg-text-primary);
+    border: 1px solid var(--sg-accent-secondary);
+    border-radius: 4px;
+    font-size: 1em; /* Consistent font size */
+    text-align: left;
+    cursor: pointer;
+    transition: background-color var(--sg-transition-fast), border-color var(--sg-transition-fast);
+}
+
+#hud-popup-menu select {
+    appearance: none; /* Basic reset for select */
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='16' height='16' fill='%23a0a0b0' viewBox='0 0 16 16'%3E%3Cpath fill-rule='evenodd' d='M1.646 4.646a.5.5 0 0 1 .708 0L8 10.293l5.646-5.647a.5.5 0 0 1 .708.708l-6 6a.5.5 0 0 1-.708 0l-6-6a.5.5 0 0 1 0-.708z'/%3E%3C/svg%3E");
+    background-repeat: no-repeat;
+    background-position: right 8px center;
+    background-size: 1em;
+    padding-right: 25px; /* Space for arrow */
+}
+
+
+#hud-popup-menu button .label {
+    margin-left: 8px;
+}
+
+#hud-popup-menu select:hover,
+#hud-popup-menu button:hover {
+    background-color: var(--sg-bg-ui-hover);
+    border-color: var(--sg-accent-primary);
+}
+#hud-popup-menu select:focus,
+#hud-popup-menu button:focus {
+    outline: none;
+    border-color: var(--sg-accent-primary);
+    box-shadow: 0 0 0 2px rgba(var(--sg-accent-primary-rgb), 0.3);
+}
+
+
+/* Individual HUD items positioning */
 #hud-selection-info {
     bottom: 10px;
     right: 10px;
 }
 
-#hud-keyboard-shortcuts {
-    top: 10px;
-    left: 10px;
-    font-size: 1.2em;
-    padding: 3px 8px;
-}
 
-#hud-layout-settings {
-    top: 10px;
-    right: 10px;
-    font-size: 1.2em;
-    padding: 3px 8px;
-}
-
-.theme-light #hud-layer {
-}
-.theme-light #hud-layer > div,
+/* Theme light adjustments */
+.theme-light #hud-layer > div:not(#hud-popup-menu),
 .theme-light #hud-layer > select {
      background-color: rgba(var(--sg-bg-ui-rgb, 255, 255, 255), 0.7);
      border: 1px solid rgba(var(--sg-accent-secondary-rgb, 0, 86, 179), 0.5);
      box-shadow: 0 1px 4px rgba(0,0,0,0.15);
 }
-.theme-light #hud-layer > div:hover,
+.theme-light #hud-layer > div:not(#hud-popup-menu):hover,
 .theme-light #hud-layer > select:hover {
     opacity: 1;
     background-color: rgba(var(--sg-bg-ui-rgb, 255, 255, 255), 0.85);
     border-color: rgba(var(--sg-accent-primary-rgb, 0, 123, 255), 0.7);
 }
+.theme-light #hud-main-menu-button.active {
+    background-color: rgba(var(--sg-accent-primary-rgb, 0, 123, 255), 0.2);
+    border-color: rgba(var(--sg-accent-primary-rgb, 0, 123, 255), 0.9);
+}
+
+.theme-light #hud-popup-menu {
+    background-color: rgba(var(--sg-bg-ui-rgb, 250, 250, 250), 0.95);
+    border-color: var(--sg-accent-primary);
+}
+.theme-light #hud-popup-menu select,
+.theme-light #hud-popup-menu button {
+    background-color: var(--sg-node-button-bg); /* Uses theme's node button bg */
+    border-color: var(--sg-accent-secondary);
+    color: var(--sg-text-primary);
+}
+.theme-light #hud-popup-menu select {
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='16' height='16' fill='%23555555' viewBox='0 0 16 16'%3E%3Cpath fill-rule='evenodd' d='M1.646 4.646a.5.5 0 0 1 .708 0L8 10.293l5.646-5.647a.5.5 0 0 1 .708.708l-6 6a.5.5 0 0 1-.708 0l-6-6a.5.5 0 0 1 0-.708z'/%3E%3C/svg%3E");
+}
+
+.theme-light #hud-popup-menu select:hover,
+.theme-light #hud-popup-menu button:hover {
+    background-color: var(--sg-bg-ui-hover); /* Uses theme's general hover */
+    border-color: var(--sg-accent-primary);
+}
+
 
 .node-label-3d,
 .edge-label {
